requirements.txt
boto==2.48.0
brotlipy==0.7.0
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
dj-database-url==0.4.2
Django==1.10
django-autoslug==1.9.3
django-bmemcached==0.2.3
django-mobile==0.7.0
django-storages==1.6.5
djangorestframework==3.7.1
djangorestframework-camel-case==0.2.0
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
newrelic==2.94.0.79
olefile==0.44
phonenumbers==8.8.3
Pillow==4.2.1
psycopg2==2.7.3.1
pycparser==2.18
PyJWT==1.5.3
PySocks==1.6.7
python-binary-memcached==0.26.1
pytz==2017.2
requests==2.18.4
six==1.11.0
twilio==6.7.0
urllib3==1.22
whitenoise==3.3.0
